Move classes to testutils to help 3rd party plugin tests
as proposed in
https://github.com/PyCQA/pylint-django/pull/252#issuecomment-547931027
