+++ ChangeLog
@@ -7,6 +7,11 @@ What's New in Pylint 2.5.0?
 
 Release date: TBA
 
+* Move ``NoFileError``, ``OutputLine``, ``FunctionalTestReporter``,
+  ``FunctionalTestFile``, ``LintModuleTest`` and related methods from
+  ``test_functional.py`` to ``pylint.testutils`` to help testing for 3rd
+  party pylint plugins.
+
 * Can read config from a setup.cfg or pyproject.toml file.
 
   Close #617
